event_api_backend/requirements.txt
# Main dependencies for Flask event management API
Flask==3.1.0
flask-smorest==0.45.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
marshmallow==3.26.1
python-dotenv==1.0.1
apispec==6.8.1
webargs==8.6.0
# Dev/test/utility
pytest==8.3.5
flake8==7.2.0
blinker==1.9.0
click==8.1.8
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
Werkzeug==3.1.3
